pyproject.toml
[tool.poetry]
name = "gotovladi"
version = "0.1.0"
description = ""
authors = ["rSlow <redmaxx4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.12.0"
aiogram-dialog = "^2.2.0"
pyyaml = "^6.0.2"
dishka = "^1.3.0"
sqlalchemy = "^2.0.32"
redis = "^5.0.8"
jinja2 = "^3.1.4"
alembic = "^1.13.2"
apscheduler = "^3.10.4"
adaptix = "^3.0.0b7"
colorlog = "^6.8.2"
asgi-monitor = "^0.5.0"
fastapi = "^0.103.2"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
alembic-postgresql-enum = "^1.3.0"
sqlalchemy-utils = "^0.41.2"
dateutils = "^0.6.12"
phonenumbers = "^8.13.44"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"
sqlalchemy-file = "^0.6.0"
pillow = "^10.4.0"
pyjwt = "^2.9.0"
fasteners = "^0.19"
pydantic-extra-types = "^2.9.0"
flask-admin = { extras = ["export", "images", "s3", "sqlalchemy", "translation"], version = "^1.6.1" }
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
flask-login = "^0.6.3"
flask-babel = "^4.0.0"
libsass = "^0.23.0"
faststream = {version = "0.5.11", extras = ["rabbit", "cli", "watchfiles"]}
pika = "^1.3.2"
watchfiles = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
